+++ src/ai_marketplace_monitor/facebook.py
@@ -73,7 +73,7 @@ class FacebookMarketItemCommonConfig(DataClassWithHandleFunc):
         if not isinstance(self.seller_locations, list) or not all(
             isinstance(x, str) for x in self.seller_locations
         ):
-            raise ValueError(f"Item {hilight(self.name, "name")} seller_locations must be a list.")
+            raise ValueError(f"Item {hilight(self.name)} seller_locations must be a list.")
 
     def handle_availability(self: "FacebookMarketItemCommonConfig") -> None:
         if self.availability is None:
@@ -82,7 +82,7 @@ class FacebookMarketItemCommonConfig(DataClassWithHandleFunc):
             x.value for x in Availability
         ]:
             raise ValueError(
-                f"Item {hilight(self.name, "name")} availability must be one of 'in' and 'out'."
+                f"Item {hilight(self.name)} availability must be one of 'in' and 'out'."
             )
 
     def handle_condition(self: "FacebookMarketItemCommonConfig") -> None:
@@ -94,7 +94,7 @@ class FacebookMarketItemCommonConfig(DataClassWithHandleFunc):
             isinstance(x, str) and x in [cond.value for cond in Condition] for x in self.condition
         ):
             raise ValueError(
-                f"Item {hilight(self.name, "name")} condition must be one or more of that can be one of 'new', 'used_like_new', 'used_good', 'used_fair'."
+                f"Item {hilight(self.name)} condition must be one or more of that can be one of 'new', 'used_like_new', 'used_good', 'used_fair'."
             )
 
     def handle_date_listed(self: "FacebookMarketItemCommonConfig") -> None:
@@ -103,9 +103,7 @@ class FacebookMarketItemCommonConfig(DataClassWithHandleFunc):
         if not isinstance(self.date_listed, int) or self.date_listed not in [
             x.value for x in DateListed
         ]:
-            raise ValueError(
-                f"Item {hilight(self.name, "name")} date_listed must be one of 1, 7, and 30."
-            )
+            raise ValueError(f"Item {hilight(self.name)} date_listed must be one of 1, 7, and 30.")
 
     def handle_delivery_method(self: "FacebookMarketItemCommonConfig") -> None:
         if self.delivery_method is None:
@@ -114,7 +112,7 @@ class FacebookMarketItemCommonConfig(DataClassWithHandleFunc):
             x.value for x in DeliveryMethod
         ]:
             raise ValueError(
-                f"Item {hilight(self.name, "name")} delivery_method must be one of 'local_pick_up' and 'shipping'."
+                f"Item {hilight(self.name)} delivery_method must be one of 'local_pick_up' and 'shipping'."
             )
 
 
@@ -319,7 +317,7 @@ class FacebookMarketplace(Marketplace):
         exclude_keywords = item_config.exclude_keywords
         if exclude_keywords and is_substring(exclude_keywords, item.title):
             self.logger.info(
-                f"""Exclude {hilight(item.title, "name")} due to {hilight("excluded keywords", "fail")}: {', '.join(exclude_keywords)}"""
+                f"""Exclude {hilight(item.title)} due to {hilight("excluded keywords", "fail")}: {', '.join(exclude_keywords)}"""
             )
             return False
 
@@ -327,7 +325,7 @@ class FacebookMarketplace(Marketplace):
         search_words = [word for keywords in item_config.keywords for word in keywords.split()]
         if not is_substring(search_words, item.title):
             self.logger.info(
-                f"""Exclude {hilight(item.title, "name")} {hilight("without search word", "fail")} in title."""
+                f"""Exclude {hilight(item.title)} {hilight("without search word", "fail")} in title."""
             )
             return False
 
@@ -335,7 +333,7 @@ class FacebookMarketplace(Marketplace):
         allowed_locations = item_config.seller_locations or self.config.seller_locations or []
         if allowed_locations and not is_substring(allowed_locations, item.location):
             self.logger.info(
-                f"""Exclude {hilight("out of area", "fail")} item {hilight(item.title, "name")} from location {hilight(item.location, "name")}"""
+                f"""Exclude {hilight("out of area", "fail")} item {hilight(item.title)} from location {hilight(item.location)}"""
             )
             return False
 
@@ -348,7 +346,7 @@ class FacebookMarketplace(Marketplace):
             and is_substring(exclude_by_description, item.description)
         ):
             self.logger.info(
-                f"""Exclude {hilight(item.title, "name")} by {hilight("description", "fail")}.\n{hilight(item.description[:100], "name")}..."""
+                f"""Exclude {hilight(item.title)} by {hilight("description", "fail")}.\n{hilight(item.description[:100])}..."""
             )
             return False
 
@@ -356,7 +354,7 @@ class FacebookMarketplace(Marketplace):
         exclude_sellers = item_config.exclude_sellers or self.config.exclude_sellers or []
         if item.seller and exclude_sellers and is_substring(exclude_sellers, item.seller):
             self.logger.info(
-                f"Exclude {hilight(item.title, "name")} sold by {hilight("banned seller", "failed")} {hilight(item.seller, "name")}"
+                f"Exclude {hilight(item.title)} sold by {hilight("banned seller", "failed")} {hilight(item.seller)}"
             )
             return False
 
@@ -529,7 +527,7 @@ class FacebookItemPage(WebPage):
                 Please report the issue to the developer if the problem persists."""
             )
 
-        self.logger.info(f"Parsing item {hilight(title, "name")}")
+        self.logger.info(f"Parsing item {hilight(title)}")
         res = SearchedItem(
             marketplace="facebook",
             name="",